Updated tests to use frame class. Many of these tests will eventually go away once the new API firms up.
git-svn-id: https://ffmpeg-php.svn.sourceforge.net/svnroot/ffmpeg-php/trunk/ffmpeg-php@261 f31480fa-141b-0410-8f85-e1b837e7f62b

+++ tests/getFrameResampled.phpt
@@ -12,15 +12,19 @@ $frame = 70;
 $mov = new ffmpeg_movie(dirname(__FILE__) . '/test_media/test.avi');
 $img = sprintf("%s/test-%04d.png", dirname(__FILE__), $frame);
 
-$gd_image = $mov->getFrameResampled(96, 120, $frame);
-if ($gd_image) {
-    imagepng($gd_image, $img);
-    imagedestroy($gd_image);
-    // generate md5 of file (NOTE: different versions of ffmpeg may produce different
-    // md5 hashes since resampling has been changed slightly due to a fix. Need to
-    // use EXPECTREX to test for both md5 possibilities.
-    printf("ffmpeg getFrameResampled(): md5 = %s\n", md5(file_get_contents($img)));
-    unlink($img);
+$ff_frame = $mov->getFrame($frame);
+if ($ff_frame) {
+    $ff_frame->resize(96, 120);
+    $gd_image = $ff_frame->toGDImage();
+    if ($gd_image) {
+        imagepng($gd_image, $img);
+        imagedestroy($gd_image);
+        // generate md5 of file (NOTE: different versions of ffmpeg may produce different
+        // md5 hashes since resampling has been changed slightly due to a fix. Need to
+        // use EXPECTREX to test for both md5 possibilities.
+        printf("ffmpeg getFrameResampled(): md5 = %s\n", md5(file_get_contents($img)));
+        unlink($img);
+    }
 }
 ?>
 --EXPECT--
